requirements.txt
numpy~=2.2.4
pandas~=2.2.3
Pillow~=11.1.0
streamlit~=1.48.0
tqdm~=4.67.1
# google service
google_api_python_client~=2.179.0

# excel
openpyxl
xlsxwriter

wheels/streamlit_drawable_canvas-0.9.3-py3-none-any.whl